Try to improve the grap a bit
The problem is that I am not even sure how to call this graph.  It
represents the instantaneous increase of value per month, that is, the
amount the account would get in 30 days at the current increase rate.

To me it makes sense cause this is generally what I am interested: not
the actual amount I did each month, but the average of how much I make
in a dt time, represented as a monthly increase, pretty much like the
instantaneous speed of the value in $/month.